requirements.txt
Django==1.6.5
South==1.0
djangorestframework==2.3.14
wsgiref==0.1.2
